remove old schedules code, more responsive changes added
