+++ grafana-plugin/src/components/SchedulesFilters/SchedulesFilters.types.ts
@@ -1,3 +1,7 @@
+import { ScheduleType } from 'models/schedule/schedule.types';
+
 export interface SchedulesFiltersType {
-  selectedDate: string;
+  searchTerm: string;
+  type: ScheduleType;
+  status: string;
 }
