--- grafana-plugin/src/components/SchedulesFilters_NEW/SchedulesFilters.module.css
@@ -1,4 +1,0 @@
-.root {
-  display: inline-flex;
-  align-items: center;
-}
