--- grafana-plugin/src/components/SchedulesFilters_NEW/SchedulesFilters.types.ts
@@ -1,7 +1,0 @@
-import { ScheduleType } from 'models/schedule/schedule.types';
-
-export interface SchedulesFiltersType {
-  searchTerm: string;
-  type: ScheduleType;
-  status: string;
-}
