+++ grafana-plugin/src/pages/schedules/Schedules.tsx
@@ -12,8 +12,8 @@ import NewScheduleSelector from 'components/NewScheduleSelector/NewScheduleSelec
 import PluginLink from 'components/PluginLink/PluginLink';
 import ScheduleCounter from 'components/ScheduleCounter/ScheduleCounter';
 import ScheduleWarning from 'components/ScheduleWarning/ScheduleWarning';
-import SchedulesFilters from 'components/SchedulesFilters_NEW/SchedulesFilters';
-import { SchedulesFiltersType } from 'components/SchedulesFilters_NEW/SchedulesFilters.types';
+import SchedulesFilters from 'components/SchedulesFilters/SchedulesFilters';
+import { SchedulesFiltersType } from 'components/SchedulesFilters/SchedulesFilters.types';
 import Table from 'components/Table/Table';
 import Text from 'components/Text/Text';
 import TimelineMarks from 'components/TimelineMarks/TimelineMarks';
@@ -29,9 +29,10 @@ import { getStartOfWeek } from 'pages/schedule/Schedule.helpers';
 import { WithStoreProps } from 'state/types';
 import { withMobXProviderContext } from 'state/withStore';
 import { UserActions } from 'utils/authorization';
-import { PLUGIN_ROOT } from 'utils/consts';
+import { PLUGIN_ROOT, TABLE_COLUMN_MAX_WIDTH } from 'utils/consts';
 
 import styles from './Schedules.module.css';
+import { MatchMediaTooltip } from 'components/MatchMediaTooltip/MatchMediaTooltip';
 
 const cx = cn.bind(styles);
 
@@ -137,9 +138,11 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
       <>
         <div className={cx('root')}>
           <VerticalGroup>
-            <HorizontalGroup justify="space-between">
-              <SchedulesFilters value={filters} onChange={this.handleSchedulesFiltersChange} />
-              <HorizontalGroup spacing="lg">
+            <div className={cx('schedules__filters-container')}>
+              <div className={cx('schedules__filters-container-left')}>
+                <SchedulesFilters value={filters} onChange={this.handleSchedulesFiltersChange} />
+              </div>
+              <div className={cx('schedules__filters-container-right')}>
                 {users && (
                   <UserTimezoneSelect
                     value={store.currentTimezone}
@@ -152,8 +155,8 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
                     + New schedule
                   </Button>
                 </WithPermissionControl>
-              </HorizontalGroup>
-            </HorizontalGroup>
+              </div>
+            </div>
             <Table
               columns={columns}
               data={data}
@@ -330,18 +333,20 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
   renderOncallNow = (item: Schedule, _index: number) => {
     if (item.on_call_now?.length > 0) {
       return (
-        <VerticalGroup>
-          {item.on_call_now.map((user, _index) => {
-            return (
-              <PluginLink key={user.pk} query={{ page: 'users', id: user.pk }}>
-                <div>
+        <div className="table__email-column">
+          <VerticalGroup>
+            {item.on_call_now.map((user, _index) => {
+              return (
+                <PluginLink key={user.pk} query={{ page: 'users', id: user.pk }} className="table__email-content">
                   <Avatar size="big" src={user.avatar} />
-                  <Text type="secondary"> {user.username}</Text>
-                </div>
-              </PluginLink>
-            );
-          })}
-        </VerticalGroup>
+                  <MatchMediaTooltip placement="top" content={user.username} maxWidth={TABLE_COLUMN_MAX_WIDTH}>
+                    <Text type="secondary"> {user.username}</Text>
+                  </MatchMediaTooltip>
+                </PluginLink>
+              );
+            })}
+          </VerticalGroup>
+        </div>
       );
     }
     return null;
